Uni TDS v1.0 - Completed Build
Amended all mistakes. Game should be functioning as expected

+++ University_TDS/Assets/Scripts/Menus/Loadout/TankBodySelection.cs
@@ -6,17 +6,17 @@ public class TankBodySelection : MonoBehaviour
     public Sprite tankBody01;
     public string tankBodyName01;
     public float tankSpeedStat01;
-    public float tankHealthStat01;
+    public int tankHealthStat01;
 
     [Header("Tank Body 02")]
     public Sprite tankBody02;
     public string tankBodyName02;
     public float tankSpeedStat02;
-    public float tankHealthStat02;
+    public int tankHealthStat02;
 
     [Header("Tank Body 03")]
     public Sprite tankBody03;
     public string tankBodyName03;
     public float tankSpeedStat03;
-    public float tankHealthStat03;
+    public int tankHealthStat03;
 }
